// src/infrastructure/external/fpl/schemas/bootstrap/element.schema.ts
import { z } from 'zod';

export const ElementResponseSchema = z
  .object({
    id: z.number(),
    code: z.number(),
    element_type: z.number(),
    first_name: z.string().nullable(),
    second_name: z.string().nullable(),
    web_name: z.string(),
    team: z.number(),
    now_cost: z.number(),
    cost_change_start: z.number(),
    team_code: z.number(),
    status: z.string(),
    minutes: z.number().default(0),
    goals_scored: z.number().default(0),
    assists: z.number().default(0),
    clean_sheets: z.number().default(0),
    goals_conceded: z.number().default(0),
    own_goals: z.number().default(0),
    penalties_saved: z.number().default(0),
    yellow_cards: z.number().default(0),
    red_cards: z.number().default(0),
    saves: z.number().default(0),
    bonus: z.number().default(0),
    bps: z.number().default(0),
    form: z.string().nullable(),
    influence: z.string().nullable(),
    creativity: z.string().nullable(),
    threat: z.string().nullable(),
    ict_index: z.string().nullable(),
    starts: z.number().default(0),
    expected_goals: z.string().nullable(),
    expected_assists: z.string().nullable(),
    expected_goal_involvements: z.string().nullable(),
    expected_goals_conceded: z.string().nullable(),
    influence_rank: z.number().nullable(),
    influence_rank_type: z.number().nullable(),
    creativity_rank: z.number().nullable(),
    creativity_rank_type: z.number().nullable(),
    threat_rank: z.number().nullable(),
    threat_rank_type: z.number().nullable(),
    ict_index_rank: z.number().nullable(),
    ict_index_rank_type: z.number().nullable(),
    total_points: z.number().default(0),
    event_points: z.number().default(0),
    cost_change_event: z.number().default(0),
    mng_win: z.number().default(0),
    mng_draw: z.number().default(0),
    mng_loss: z.number().default(0),
    mng_underdog_win: z.number().default(0),
    mng_underdog_draw: z.number().default(0),
    mng_clean_sheets: z.number().default(0),
    mng_goals_scored: z.number().default(0),
    value_form: z.string().nullable(),
    value_season: z.string().nullable(),
    transfers_in: z.number().default(0),
    transfers_out: z.number().default(0),
    transfers_in_event: z.number().default(0),
    transfers_out_event: z.number().default(0),
    chance_of_playing_this_round: z.number().nullable(),
    chance_of_playing_next_round: z.number().nullable(),
    selected_by_percent: z.string().nullable(),
  })
  .passthrough();

export type ElementResponse = z.infer<typeof ElementResponseSchema>;